aero.cpp now computes partial flow for when the airflow is cutoff by other surfaces (only for the rear wing). Modified the flow cut off for when you are following another car. When this happens both your drag and your downforce is decreased. When a car is behind you, only the eddies contributing to drag are reduced, so in that case I only reduce the drag. Did the same for the wings, taking their position into account (so the front wing will have its airflow reduced more than the rear). Incidentally fixed a bug in the calculation (dragK=dragK*tmpas instead of dragK=tmpas, since tmpas<1 anyway). Complaints that the cars are difficult to drive in curves do not seem to apply, at least not for the high-speed banked tracks now that the full 3d calculation is in effect.
car.cpp now calculates forces in full 3d. Note that in order to apply
accelerations we need to rotate from the car axis to the world axis,
so I apply NaiveInverseRotate() there.

wheel.cpp now calculates forces in full 3d. However, there is one caveat
related to the suspension and reaction forces. Normally there are reaction forces to the two axis that are perpendicular to the suspension as well. Currently
I am approximating those in a slightly hacky way. However, tests using an exaggerated version of Dirt-2 show that the model works OK.

Other caveats: When the car is upside down forces become unbounded and the
simulation fails catastrophically.